@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Windows.Forms;
-
-namespace Aim_2_MoTeC
+﻿namespace Aim_2_MoTeC
 {
     class MotecLog
     {
