@@ -1,7 +1,4 @@
-﻿using System;
-using System.Diagnostics;
-using System.Runtime.InteropServices;
-using System.Windows.Forms;
+﻿using System.Runtime.InteropServices;
 
 namespace Aim_2_MoTeC
 {
@@ -21,76 +18,83 @@ namespace Aim_2_MoTeC
     public static class XRK
     {
         /// <summary>Name of the library</summary>
-        const string libName = "MatLabXRK-2017-64-ReleaseU.dll";
+        const string libName = "MatLabXRK-2022-64-ReleaseU.dll";
+        const CallingConvention callingConvention = CallingConvention.Cdecl;
 
         /// <summary>Get the compile date of this library</summary>
         /// <returns>Compile date</returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_library_date();
 
         /// <summary>Get the compile time of this library </summary>
         /// <returns>Compile time</returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_library_time();
 
         /// <summary>Open a xrk file</summary>
         /// <param name="full_path_name">full path to the file to be opened, it cannot be a relative path</param>
         /// <returns>graeter than 0 in case of success the internal index of the file opened, 0 in case the file is opened but can't be parsed, less than 0 in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Ansi)]
+        [DllImport(libName, CallingConvention = callingConvention, CharSet = CharSet.Ansi)]
         private static extern int open_file(string full_path_name);
+
+        /// <summary>Open a xrk file with a licence</summary>
+        /// <param name="full_path_name">full path to the file to be opened, it cannot be a relative path</param>
+        /// <returns>graeter than 0 in case of success the internal index of the file opened, 0 in case the file is opened but can't be parsed, less than 0 in case of problems </returns>
+        [DllImport(libName, CallingConvention = callingConvention, CharSet = CharSet.Ansi)]
+        private static extern int open_file_with_licence(string full_path_name, string full_path_licence);
 
         /// <summary>Close a xrk file</summary>
         /// <param name="full_path_name">full path to the file to be closed, it cannot be a relative path</param>
         /// <returns> greater than 0 the internal index of the file closed, less than 0 in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Ansi)]
+        [DllImport(libName, CallingConvention = callingConvention, CharSet = CharSet.Ansi)]
         private static extern int close_file_n(string full_path_name);
 
         /// <summary>Close a xrk file</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> greater than 0 the internal index of the file closed, less than 0 in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int close_file_i(int idx);
 
         /// <summary>Get vehicle info</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_vehicle_name(int idx);
 
         /// <summary>Get track info</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_track_name(int idx);
 
         /// <summary>Get racer info</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_racer_name(int idx);
 
         /// <summary>Get championship info</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_championship_name(int idx);
 
         /// <summary>Get venue type info</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
-        private static extern IntPtr get_venue_type_name(int idx);
+        [DllImport(libName, CallingConvention = callingConvention)]
+        private static extern IntPtr get_session_type_name(int idx);
 
         /// <summary>Get session date and time</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> a struct tm pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_date_and_time(int idx);
 
         /// <summary>Get laps count of a xrk file</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> greater than 0 the laps count, less than 0 in case of problems, 0 in case the xrk file has no laps (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_laps_count(int idx);
 
         /// <summary>Get lap info</summary>
@@ -99,34 +103,34 @@ namespace Aim_2_MoTeC
         /// <param name="pstart">pointer to an already allocated double in which the start time of the lap will be stored</param>
         /// <param name="pduration">pointer to an already allocated double in which the lap duration will be stored</param>
         /// <returns> 1 if all is OK, less than 0 in case of problems, 0 in case the xrk file has no laps (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_lap_info(int idxf, int idxl, out double pstart, out double pduration);
 
         /// <summary> Get channels count of a xrk file</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns> 1 if all is OK, less than 0 in case of problems, 0 in case the xrk file has no laps (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_channels_count(int idx);
 
         /// <summary>Get channel name</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_channel_name(int idxf, int idxc);
 
         /// <summary>Get channel units</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_channel_units(int idxf, int idxc);
 
         /// <summary>Get channel samples count</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_channel_samples_count(int idxf, int idxc);
 
         /// <summary>Get channel samples</summary>
@@ -136,7 +140,7 @@ namespace Aim_2_MoTeC
         /// <param name="pvalues">pointer to an already allocated buffer in which sample values will be stored</param>
         /// <param name="cnt">the number of samples of the buffer, that is the samples count returned by get channel samples count function</param>
         /// <returns>greater than 0 the samples count, less than 0 in case of problems, 0 in case the cnt param doesn't match the samples count or in case the channel has no samples (theorically not possible)</returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_channel_samples(int idxf, int idxc, double[] ptimes, double[] pvalues, int cnt);
 
         /// <summary>Get channel samples count in a given lap</summary>
@@ -144,7 +148,7 @@ namespace Aim_2_MoTeC
         /// <param name="idxl">the lap index</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_lap_channel_samples_count(int idxf, int idxl, int idxc);
 
         /// <summary>Get channel samples in a given lap</summary>
@@ -155,34 +159,34 @@ namespace Aim_2_MoTeC
         /// <param name="pvalues">pointer to an already allocated buffer in which sample values will be stored</param>
         /// <param name="cnt">the number of samples of the buffer, that is the samples count returned by get channel samples count function</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the cnt param doesn't match the samples count or in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_lap_channel_samples(int idxf, int idxl, int idxc, double[] ptimes, double[] pvalues, int cnt);
 
         /// <summary> Get GPS channels count of a xrk file</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns>greater than 0 the channel count, less than 0 in case of problems, 0 in case the xrk file has no channels (theorically not possible)</returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_GPS_channels_count(int idx);
 
         /// <summary>Get GPS channel name</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_GPS_channel_name(int idxf, int idxc);
 
         /// <summary>Get GPS channel units</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_GPS_channel_units(int idxf, int idxc);
 
         /// <summary>Get GPS channel samples count</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_GPS_channel_samples_count(int idxf, int idxc);
 
         /// <summary>Get GPS channel samples</summary>
@@ -192,7 +196,7 @@ namespace Aim_2_MoTeC
         /// <param name="pvalues">pointer to an already allocated buffer in which sample values will be stored</param>
         /// <param name="cnt">the number of samples of the buffer, that is the samples count returned by get channel samples count function</param>
         /// <returns>greater than 0 the samples count, less than 0 in case of problems, 0 in case the cnt param doesn't match the samples count or in case the channel has no samples (theorically not possible)</returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_GPS_channel_samples(int idxf, int idxc, double[] ptimes, double[] pvalues, int cnt);
 
         /// <summary>Get GPS channel samples count in a given lap</summary>
@@ -200,7 +204,7 @@ namespace Aim_2_MoTeC
         /// <param name="idxl">the lap index</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_lap_GPS_channel_samples_count(int idxf, int idxl, int idxc);
 
         /// <summary>Get GPS channel samples in a given lap</summary>
@@ -211,34 +215,34 @@ namespace Aim_2_MoTeC
         /// <param name="pvalues">pointer to an already allocated buffer in which sample values will be stored</param>
         /// <param name="cnt">the number of samples of the buffer, that is the samples count returned by get channel samples count function</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the cnt param doesn't match the samples count or in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_lap_GPS_channel_samples(int idxf, int idxl, int idxc, double[] ptimes, double[] pvalues, int cnt);
 
         /// <summary> Get GPS raw channels count of a xrk file</summary>
         /// <param name="idx">the internal file index returned by open function</param>
         /// <returns>greater than 0 the channel count, less than 0 in case of problems, 0 in case the xrk file has no channels (theorically not possible)</returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_GPS_raw_channels_count(int idx);
 
         /// <summary>Get GPS raw channel name</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_GPS_raw_channel_name(int idxf, int idxc);
 
         /// <summary>Get GPS raw channel units</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> a null terminated char pointer with the required info, NULL in case of problems </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern IntPtr get_GPS_raw_channel_units(int idxf, int idxc);
 
         /// <summary>Get GPS raw channel samples count</summary>
         /// <param name="idxf">the internal file index returned by open function</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_GPS_raw_channel_samples_count(int idxf, int idxc);
 
         /// <summary>Get GPS raw channel samples</summary>
@@ -248,7 +252,7 @@ namespace Aim_2_MoTeC
         /// <param name="pvalues">pointer to an already allocated buffer in which sample values will be stored</param>
         /// <param name="cnt">the number of samples of the buffer, that is the samples count returned by get channel samples count function</param>
         /// <returns>greater than 0 the samples count, less than 0 in case of problems, 0 in case the cnt param doesn't match the samples count or in case the channel has no samples (theorically not possible)</returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_GPS_raw_channel_samples(int idxf, int idxc, double[] ptimes, double[] pvalues, int cnt);
 
         /// <summary>Get GPS raw channel samples count in a given lap</summary>
@@ -256,7 +260,7 @@ namespace Aim_2_MoTeC
         /// <param name="idxl">the lap index</param>
         /// <param name="idxc">the channel index</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_lap_GPS_raw_channel_samples_count(int idxf, int idxl, int idxc);
 
         /// <summary>Get GPS raw channel samples in a given lap</summary>
@@ -267,7 +271,7 @@ namespace Aim_2_MoTeC
         /// <param name="pvalues">pointer to an already allocated buffer in which sample values will be stored</param>
         /// <param name="cnt">the number of samples of the buffer, that is the samples count returned by get channel samples count function</param>
         /// <returns> greater than 0 the samples count, less than 0 in case of problems, 0 in case the cnt param doesn't match the samples count or in case the channel has no samples (theorically not possible) </returns>
-        [DllImport(libName, CallingConvention = CallingConvention.Cdecl)]
+        [DllImport(libName, CallingConvention = callingConvention)]
         private static extern int get_lap_GPS_raw_channel_samples(int idxf, int idxl, int idxc, double[] ptimes, double[] pvalues, int cnt);
 
 
@@ -342,7 +346,7 @@ namespace Aim_2_MoTeC
         /// <returns>the venue type </returns>
         public static string GetVenueTypeName(int idx)
         {
-            IntPtr namePtr = get_venue_type_name(idx);
+            IntPtr namePtr = get_session_type_name(idx);
             string name = Marshal.PtrToStringAnsi(namePtr);
             return name;
         }
@@ -579,14 +583,29 @@ namespace Aim_2_MoTeC
         {
             try
             {
+                // 0 in case the file is opened but can't be parsed, less than 0 in case of problems
                 id = OpenFile(path);
-                return true;
+                if (id > 0)
+                {
+                    return true;
+                }
+                else if (id == 0)
+                {
+                    MessageBox.Show($"File can't be parsed using {libName}", "Error", MessageBoxButtons.OK);
+                    return false;
+                }
+                else //id < 0
+                {
+                    MessageBox.Show($"Problem happended while parsing ", "Error", MessageBoxButtons.OK);
+                    return false;
+                }
+
             }
             catch (Exception ex)
             {
-                MessageBox.Show("Failed to load MatLabXRK-2017-64-ReleaseU.dll", "Error", MessageBoxButtons.OK);
+                MessageBox.Show($"Failed to load {libName}", "Error", MessageBoxButtons.OK);
 
-                Console.WriteLine("Failed to load MatLabXRK-2017-64-ReleaseU.dll");
+                Console.WriteLine($"Failed to load {libName}");
                 Console.WriteLine(ex.Message);
                 id = -1;
                 return false;
