@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
+﻿using System.Text;
 namespace Aim_2_MoTeC
 {
     using static Encoder;
